% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_Arabidopsis.thaliana.Araport11.R
\docType{data}
\name{rvhys_Arabidopsis.thaliana.Araport11.Gene.Description}
\alias{rvhys_Arabidopsis.thaliana.Araport11.Gene.Description}
\title{Arabidopsis thaliana gene description data}
\format{
\subsection{\code{tibble}}{

A tibble with 37,897 rows and 6 columns:
}
}
\source{
\url{https://www.arabidopsis.org/tools/bulk/genes/index.jsp}
}
\usage{
rvhys_Arabidopsis.thaliana.Araport11.Gene.Description
}
\description{
Arabidopsis thaliana gene description data
}
\keyword{datasets}
