% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_getDNAseq.R
\name{rvhys_getDNAseq}
\alias{rvhys_getDNAseq}
\title{extract DNA sequence from Chr, start, end and strand}
\usage{
rvhys_getDNAseq(chr, start, end, strand)
}
\description{
extract DNA sequence from Chr, start, end and strand
}
\examples{
rvhys_getDNAseq(3, 19329976, 19330011, "+")

}
