% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_intronID2seq.R
\name{rvhys_intronID2seq}
\alias{rvhys_intronID2seq}
\title{根据intron ID提取DNA序列}
\usage{
rvhys_intronID2seq(intron)
}
\arguments{
\item{intron}{}
}
\description{
根据intron ID提取DNA序列
}
\examples{
rvhys_intronID2seq("AT4G25590.1intron1")

}
