% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_na_outliers.R
\name{rvhys_na_outliers}
\alias{rvhys_na_outliers}
\title{Replace outliers with NA}
\usage{
rvhys_na_outliers(data, var)
}
\arguments{
\item{data}{}

\item{var}{}
}
\description{
Replace outliers with NA
}
\examples{
x <- tibble::tibble(a = c(1, 30, 33, 37), b = c(15, 16, 18, 100))
rvhys_na_outliers(x, a)
rvhys_na_outliers(x, b)

}
