% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_perm.R
\name{rvhys_perm}
\alias{rvhys_perm}
\title{convenient permutation test}
\usage{
rvhys_perm(x, y, reps = 10000, seed = 20230425)
}
\description{
convenient permutation test
}
\examples{
a <- runif(5, 1, 2)
b <- runif(6, 4, 5)
rvhys_perm(a, b)

}
