% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_perm2.R
\name{rvhys_perm2}
\alias{rvhys_perm2}
\title{pipe-friendly permutation test}
\usage{
rvhys_perm2(data, formula, reps = 10000)
}
\arguments{
\item{reps}{}
}
\description{
pipe-friendly permutation test
}
\examples{
tibble::tibble(a = runif(5, 1, 2), b = runif(5, 4, 5)) \%>\%
  pivot_longer(cols = everything(), names_to = "group", values_to = "value") \%>\%
  rvhys_perm2(value ~ group)
}
