% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_qPCR_primer.R
\name{rvhys_qPCR_primer}
\alias{rvhys_qPCR_primer}
\title{Pick qPCR primer}
\source{
\url{https://biodb.swu.edu.cn/qprimerdb/}
}
\usage{
rvhys_qPCR_primer(gene)
}
\arguments{
\item{gene}{}
}
\description{
Pick qPCR primer
}
\examples{
rvhys_qPCR_primer("At4g31770")

}
