% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_rescale.R
\name{rvhys_rescale}
\alias{rvhys_rescale}
\title{归一化}
\usage{
rvhys_rescale(x, type = "pos", min = 0, max = 1)
}
\arguments{
\item{type}{}
}
\description{
min-max normalization
}
\examples{
rvhys_rescale(c(1:5, 7:22, Inf), min = 1, max = 2)

}
