% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_theme_1.R
\name{rvhys_theme_1}
\alias{rvhys_theme_1}
\title{Custom ggplot2 theme 1}
\usage{
rvhys_theme_1(
  base_size = 6,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_rect_size}{}
}
\description{
This function is used to custom ggplot2 theme derived from theme_classic.
}
\examples{
#' mpg \%>\%
ggplot(aes(cty, hwy)) +
  geom_point() +
  facet_wrap(~class) +
  rvhys_theme_1()

}
