% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_theme_2.R
\name{rvhys_theme_2}
\alias{rvhys_theme_2}
\title{Custom ggplot2 theme}
\usage{
rvhys_theme_2(
  base_size = 6,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_rect_size}{}
}
\description{
This function is used to custom ggplot2 theme derived from theme_test.
}
