% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvhys_write_xlsx.R
\name{rvhys_write_xlsx}
\alias{rvhys_write_xlsx}
\title{export formatted excel}
\usage{
rvhys_write_xlsx(df, file)
}
\description{
导出适合学术发表的excel表格风格，后续还需要手动调整列宽，单元格对其方式，如有
必要，也可以修改窗口冻结设置
}
\examples{
\donttest{
rvhys_write_xlsx(iris, "test.xlsx")
}

}
